FCOM 1.4 to 1.5
0.1.5
-Added: More descriptions in the settings and tooltips of settings
-Added on request: Randomize mounts for zone & subZone
Enable the randomize mounts checkbox.
Choose your zone & subZone, choose your mount from the dropdown of mounts.
Click the + button to add them to the randomized zone & subzone dropdown
Select one entry in this dropdown and click the - button to remove the mount again.

If you enable the -ALL-: Use the -ALL- subzone for the mounts, this will also apply to the random mounts!
This means if you enable this setting you need to select the zone and the subzone -ALL- to add (+) or delete (-) the random mounts to the dropdown box.

Random mounts will be selected the same like normal mounts, a few seconds after unmounting or at a zone change or /reloadui then!

0.1.4
-Updated API for Drgaonhold
-Added zone entries for Southern Elsweyr
-Removed unregister for callback of LAM panel created