FCOMounty 0.1.7 -Made compatible with and dependend on LibZone version 0.6.8 (mandatory!) -Updated API to Greymoor -Added Greymoor Skyrim and underground cavern zone data -Fixed map tile texture detection to support (hopefully all) map texture names now (thanks to manavortex, Shalikran, sirinsidiator and kyoma) -Added chat slash command /fcomz to show the current zone info in the chat -Added chat slash command /fcomr to report a missing zone or subzone directly to me. This will only work if you are not in a house or dungeon, if you are on an overland map or a city map (not all cities provide "subzone" names so maybe some cities are named the same in their zone as the overland zone. I cannot change this sorry. WAtch the chat for information if you are standing wrong, the data already exists or not.
If the zone or subzone is not known yet in the FCOMounty data tables:
A popup will open, trying to open a website www.esoui.com and so on... Click yes and it will auto fill in the needed zone data for me into a bug report at www.esoui.com -> FCOMounty bug panel. Send this bug report then and I got all data to add.